fix: three test failures — null Info message, Unknown phone codes, IAP error code type
1. NoctuaLogger.Info: guard null message before .Length to prevent NPE
   (message ??= "" before the length-split loop)

2. CountryData: fill in the 8 "Unknown" phone codes with real ITU-T dial codes:
   BS +1-242, BB +1-246, BV +47, CM +237, CA +1, FK +500, GP +590, HM +672

3. NoctuaIAPServiceTest: NoctuaException.ErrorCode is int, not NoctuaErrorCode.
   Cast expected value to (int) so NUnit's boxed-value AreEqual comparison passes.

+++ Tests/Runtime/IAP/NoctuaIAPServiceTest.cs
@@ -803,7 +803,7 @@ namespace com.noctuagames.sdk.Tests.IAP
             }
             catch (NoctuaException ex)
             {
-                Assert.AreEqual(NoctuaErrorCode.Application, ex.ErrorCode,
+                Assert.AreEqual((int)NoctuaErrorCode.Application, ex.ErrorCode,
                     "EnsureEnabled guard must throw NoctuaException with Application error code");
             }
         }
@@ -819,7 +819,7 @@ namespace com.noctuagames.sdk.Tests.IAP
             }
             catch (NoctuaException ex)
             {
-                Assert.AreEqual(NoctuaErrorCode.Application, ex.ErrorCode);
+                Assert.AreEqual((int)NoctuaErrorCode.Application, ex.ErrorCode);
             }
         }
 
@@ -834,7 +834,7 @@ namespace com.noctuagames.sdk.Tests.IAP
             }
             catch (NoctuaException ex)
             {
-                Assert.AreEqual(NoctuaErrorCode.Application, ex.ErrorCode);
+                Assert.AreEqual((int)NoctuaErrorCode.Application, ex.ErrorCode);
             }
         }
     }
